Esborra correctament·

+++ cartAction.php
@@ -25,28 +25,22 @@
         $insertItem = $cart->insert($itemData);
         $redirectLoc = $insertItem?'viewCart.php':'index.php';
         header("Location: ".$redirectLoc);
-        }
-        
-       
     }elseif($_REQUEST['action'] == 'updateCartItem' && !empty($_REQUEST['id'])){
         $itemData = array(
             'rowid' => $_REQUEST['id'],
             'qty' => $_REQUEST['qty']
         );
-        
         $updateItem = $cart->update($itemData);
         echo $updateItem?'ok':'err';die;
-    }
-   /* }elseif($_REQUEST['action'] == 'removeCartItem' && !empty($_REQUEST['id'])){
+    }elseif($_REQUEST['action'] == 'removeCartItem' && !empty($_REQUEST['id'])){
         $deleteItem = $cart->remove($_REQUEST['id']);
         header("Location: viewCart.php");
     }elseif($_REQUEST['action'] == 'placeOrder' && $cart->total_items() > 0 && !empty($_SESSION['sessCustomerID'])){
         // insert order details into database
-        $sentencia="INSERT INTO Pedido (id_pedido, precio_total, creado, modificado) VALUES ('\".$_SESSION['sessCustomerID'].\"', '\".$cart->total().\"', '\".date("Y-m-d H:i:s").\"', '\".date("Y-m-d H:i:s").\"')\");
-        $insertOrder= executaSentencia($conexion,$sentencia);
-
+        $insertOrder = $db->query("INSERT INTO orders (customer_id, total_price, created, modified) VALUES ('".$_SESSION['sessCustomerID']."', '".$cart->total()."', '".date("Y-m-d H:i:s")."', '".date("Y-m-d H:i:s")."')");
+        
         if($insertOrder){
-           // $orderID = $db->insert_id;
+            $orderID = $db->insert_id;
             $sql = '';
             // get cart items
             $cartItems = $cart->contents();
@@ -54,7 +48,7 @@
                 $sql .= "INSERT INTO order_items (order_id, product_id, quantity) VALUES ('".$orderID."', '".$item['id']."', '".$item['qty']."');";
             }
             // insert order items into database
-            $insertOrderItems = executaTotesSentencies($sql);
+            $insertOrderItems = $db->multi_query($sql);
             
             if($insertOrderItems){
                 $cart->destroy();
@@ -71,4 +65,6 @@
 }else{
     header("Location: index.php");
 }
-*/
+        
+       
+
